+++ migrations/34_hpolygon.js
@@ -1,16 +1,12 @@
-/** Archived
 const Registry = artifacts.require('Registry');
-const Handler = artifacts.require('HMooniswap');
+const Handler = artifacts.require('HPolygon');
 const utils = web3.utils;
-*/
 
 module.exports = async function(deployer) {
-  /** Archived
   if (deployer.network === 'development') {
     return;
   }
   await deployer.deploy(Handler);
   const registry = await Registry.deployed();
-  await registry.register(Handler.address, utils.asciiToHex('HMooniswap'));
-*/
+  await registry.register(Handler.address, utils.asciiToHex('HPolygon'));
 };