+++ test/HAaveProtocolV2_flashloan.test.js
@@ -28,7 +28,12 @@ const {
   AWETH_V2_DEBT_VARIABLE,
   AAVE_RATEMODE,
 } = require('./utils/constants');
-const { evmRevert, evmSnapshot, profileGas } = require('./utils/utils');
+const {
+  evmRevert,
+  evmSnapshot,
+  profileGas,
+  errorCompare,
+} = require('./utils/utils');
 
 const Registry = artifacts.require('Registry');
 const Proxy = artifacts.require('ProxyMock');
@@ -40,7 +45,6 @@ const SimpleToken = artifacts.require('SimpleToken');
 const IToken = artifacts.require('IERC20');
 const ILendingPoolV2 = artifacts.require('ILendingPoolV2');
 const IProviderV2 = artifacts.require('ILendingPoolAddressesProviderV2');
-const IUniswapExchange = artifacts.require('IUniswapExchange');
 const IVariableDebtToken = artifacts.require('IVariableDebtToken');
 const IStableDebtToken = artifacts.require('IStableDebtToken');
 
@@ -158,9 +162,7 @@ contract('AaveV2 flashloan', function([_, user, someone]) {
         0,
         { from: this.tokenBProvider }
       );
-      expect(await this.aTokenB.balanceOf.call(user)).to.be.bignumber.eq(
-        depositAmount
-      );
+      errorCompare(await this.aTokenB.balanceOf.call(user), depositAmount);
     });
 
     it('single asset with no debt', async function() {