+++ test/HUniswapV2_liquidity.test.js
@@ -453,7 +453,7 @@ contract('UniswapV2 Liquidity', function([_, user]) {
         deadline,
         {
           from: user,
-          value: ether('0.1'),
+          value: ether('1'),
         }
       );