+++ test/utils/utils.js
@@ -2,6 +2,8 @@ const { BN, ether } = require('@openzeppelin/test-helpers');
 const fetch = require('node-fetch');
 const { ETH_PROVIDER, RecordHandlerResultSig } = require('./constants');
 
+const { expect } = require('chai');
+
 function profileGas(receipt) {
   receipt.logs.forEach(element => {
     if (element.event === 'DeltaGas')
@@ -88,6 +90,25 @@ function getHandlerReturn(receipt, dataTypes) {
   return handlerResult;
 }
 
+function errorCompare(a, b, e = new BN('1')) {
+  expect(a.sub(b).abs()).to.be.bignumber.lte(e);
+}
+
+// Only works when one function name matches
+function getAbi(artifact, name) {
+  var abi;
+  artifact.abi.forEach((element, i) => {
+    if (element.name === name) {
+      abi = element;
+    }
+  });
+  return abi;
+}
+
+function getCallData(artifact, name, params) {
+  return web3.eth.abi.encodeFunctionCall(getAbi(artifact, name), params);
+}
+
 module.exports = {
   profileGas,
   evmSnapshot,
@@ -96,4 +117,7 @@ module.exports = {
   mulPercent,
   cUnit,
   getHandlerReturn,
+  errorCompare,
+  getAbi,
+  getCallData,
 };